ASP-4668 Split License Manager Simulator
