+++ lm-simulator-api/docker-compose.yaml
@@ -43,7 +43,7 @@ services:
       DATABASE_NAME: local-db
       DATABASE_PORT: 5432
     ports: ["8000:8000"]
-    command: uvicorn lm_simulator.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
+    command: uvicorn lm_simulator_api.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
 
   unittest:
     build:
