+++ lm-simulator-api/lm_simulator_api/database.py
@@ -3,8 +3,8 @@ from typing import AsyncGenerator
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 from yarl import URL
 
-from lm_simulator.config import settings
-from lm_simulator.models import Base
+from lm_simulator_api.models import Base
+from lm_simulator_api.config import settings
 
 engine = create_async_engine(
     str(
