+++ lm-simulator/pyproject.toml
@@ -8,39 +8,13 @@ packages = [{ include = "lm_simulator" }]
 
 [tool.poetry.dependencies]
 python = "^3.12"
-fastapi = "^0.111.0"
-pydantic = "^2.8.2"
-pydantic-settings = "^2.3.4"
-SQLAlchemy = {extras = ["mypy"], version = "^2.0.22"}
-uvicorn = "^0.30.1"
-requests = "^2.27.1"
-psycopg2 = "2.9.5"
-yarl = "^1.9.2"
-asyncpg = "^0.28"
-py-buzz = "^4.1.0"
+requests = "^2.26.0"
+Jinja2 = "^3.0.1"
 
 [tool.poetry.dev-dependencies]
-pytest = "^6.2.4"
-pytest-asyncio = "^0.15.1"
-pytest-cov = "^2.12.1"
-pytest-env = "^0.6.2"
-pytest-random-order = "^1.0.4"
 black = "^24.4.2"
 flake8 = "^7.0.0"
 isort = "^5.9.3"
-requests = "^2.26.0"
-Jinja2 = "^3.0.1"
-python-dotenv = "^1.0.1"
-
-[tool.pytest.ini_options]
-addopts = "-v --random-order --cov=lm_simulator --cov-report=term-missing --cov-fail-under=85"
-env = [
-    "TEST_DATABASE_HOST: test-db",
-    "TEST_DATABASE_USER: test-user",
-    "TEST_DATABASE_PSWD: test-pswd",
-    "TEST_DATABASE_NAME: test-db",
-    "TEST_DATABASE_PORT: 5432"
-]
 
 [tool.black]
 line-length = 110
